test: Add integration test using gnomock

+++ migrations/001_init.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE files.files (
     id                    TEXT PRIMARY KEY,
     key                   TEXT NOT NULL,
     local_path            TEXT NOT NULL,
-    checksum              INTEGER NOT NULL,
+    checksum              BIGINT NOT NULL,
     etag                  TEXT NOT NULL,
     bucket                TEXT NOT NULL,
     version               TEXT NOT NULL,
